move core/types.go to its own package; update other packages
