+++ core/manager.go
@@ -1,4 +1,20 @@
-package manager
+// Copyright 2019 ChainSafe Systems (ON) Corp.
+// This file is part of gossamer.
+//
+// The gossamer library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The gossamer library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
+
+package core
 
 import (
 	log "github.com/ChainSafe/log15"
@@ -6,7 +22,7 @@ import (
 	scale "github.com/ChainSafe/gossamer/codec"
 	tx "github.com/ChainSafe/gossamer/common/transaction"
 	"github.com/ChainSafe/gossamer/consensus/babe"
-	"github.com/ChainSafe/gossamer/core"
+	"github.com/ChainSafe/gossamer/core/types"
 	"github.com/ChainSafe/gossamer/p2p"
 	"github.com/ChainSafe/gossamer/runtime"
 )
@@ -38,6 +54,8 @@ func (s *Service) start(e chan error) {
 		msgType := msg.GetType()
 		if msgType == p2p.TransactionMsgType {
 			// process tx
+		} else if msgType == p2p.BlockAnnounceMsgType {
+			// process block
 		}
 	}(s.msgChan)
 
@@ -52,7 +70,7 @@ func (s *Service) Stop() <-chan error {
 
 // ProcessTransaction attempts to validates the transaction
 // if it is validated, it is added to the transaction pool of the BABE session
-func (s *Service) ProcessTransaction(e core.Extrinsic) error {
+func (s *Service) ProcessTransaction(e types.Extrinsic) error {
 	validity, err := s.validateTransaction(e)
 	if err != nil {
 		log.Error("ProcessTransaction", "error", err)
@@ -67,13 +85,13 @@ func (s *Service) ProcessTransaction(e core.Extrinsic) error {
 
 // ProcessBlock attempts to add a block to the chain by calling `core_execute_block`
 // if the block is validated, it is stored in the block DB and becomes part of the canonical chain
-func (s *Service) ProcessBlock(b *core.BlockHeader) {
-	return
+func (s *Service) ProcessBlock(b *types.BlockHeader) error {
+	return nil
 }
 
 // runs the extrinsic through runtime function TaggedTransactionQueue_validate_transaction
 // and returns *Validity
-func (s *Service) validateTransaction(e core.Extrinsic) (*tx.Validity, error) {
+func (s *Service) validateTransaction(e types.Extrinsic) (*tx.Validity, error) {
 	ret, err := s.rt.Exec("TaggedTransactionQueue_validate_transaction", 1, 0)
 	if err != nil {
 		return nil, err
